pyproject.toml
[tool.poetry]
name = "washcar-api"
version = "0.1.0"
description = ""
authors = ["Ruhin Shukurlu <ruhinshukurlu@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "washcar_api"}]

[tool.poetry.dependencies]
python = "3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
